+++ scripts/generate_excel.py
@@ -60,7 +60,7 @@ def create_excel_file(args):
     five_utr_sheet = []
     misc_sheet = []
 
-    header = ["Identifier", "Genome", "Source", "Feature", "Start", "Stop", "Strand", "Locus_tag", "Old_locus_tag", "Name", "Length", "Codon_count"] + [cond + "_TE" for cond in TE_header] + [card + "_rpkm" for card in wildcards] + ["Start_codon", "Stop_codon", "Nucleotide_seq", "Aminoacid_seq",  "Product", "Note"]
+    header = ["Identifier", "Genome", "Source", "Feature", "Start", "Stop", "Strand", "Locus_tag", "Old_locus_tag", "Name", "Length", "Codon_count"] + [cond + "_TE" for cond in TE_header] + [card + "_rpkm" for card in wildcards] + ["Start_codon", "Stop_codon", "15nt upstream", "Nucleotide_seq", "Aminoacid_seq",  "Product", "Note"]
     prefix_columns = len(read_df.columns) - len(wildcards)
     name_list = ["s%s" % str(x) for x in range(len(header))]
     nTuple = collections.namedtuple('Pandas', name_list)
@@ -74,7 +74,7 @@ def create_excel_file(args):
         strand = getattr(row, "_6")
         attributes = getattr(row, "_8")
 
-        start_codon, stop_codon, nucleotide_seq, aa_seq = eu.get_genome_information(genome_dict[chromosome], start-1, stop-1, strand)
+        start_codon, stop_codon, nucleotide_seq, aa_seq, nt_window = eu.get_genome_information(genome_dict[chromosome], start-1, stop-1, strand)
         pred_value, name, product, note, evidence, locus_tag, old_locus_tag = eu.retrieve_column_information(attributes)
 
         length = stop - start + 1
@@ -88,7 +88,7 @@ def create_excel_file(args):
         TE_list = eu.calculate_TE(rpkm_list, wildcards, conditions)
 
         identifier = "%s:%s-%s:%s" % (chromosome, start, stop, strand)
-        result = [identifier, chromosome, "HRIBO", feature, start, stop, strand, locus_tag, old_locus_tag, name, length, codon_count] + TE_list + rpkm_list + [start_codon, stop_codon, nucleotide_seq, aa_seq, product, note]
+        result = [identifier, chromosome, "HRIBO", feature, start, stop, strand, locus_tag, old_locus_tag, name, length, codon_count] + TE_list + rpkm_list + [start_codon, stop_codon, nt_window, nucleotide_seq, aa_seq, product, note]
 
         all_sheet.append(nTuple(*result))